Finished depiction of election time-line, only left to draw it properly on start after long pause. Added some calculations around reward collection.

+++ src/main/java/org/ton/executors/liteclient/LiteClientParser.java
@@ -1611,19 +1611,19 @@ public class LiteClientParser {
 
         if (StringUtils.isEmpty(stdout) || !stdout.contains("compute_returned_stake"))
             return ResultComputeReturnStake.builder()
-                    .stake(new BigInteger("-1"))
+                    .stake(new BigDecimal("-1"))
                     .build();
 
         if (stdout.contains("cannot parse answer")) {
             return ResultComputeReturnStake.builder()
-                    .stake(new BigInteger("-1"))
+                    .stake(new BigDecimal("-1"))
                     .build();
         }
 
         String result = StringUtils.substringBetween(stdout, "result:  [", "]");
 
         return ResultComputeReturnStake.builder()
-                .stake(new BigInteger(result.trim()))
+                .stake(new BigDecimal(result.trim()))
                 .build();
     }
 }
